redux周りの構成を変更
